requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
PyYAML==6.0.1
python-telegram-bot==21.10
schedule==1.2.0
cloudscraper==1.2.71
httpx==0.28.1